+++ cpp/test/common/serde_test.cpp
@@ -13,22 +13,23 @@
 // limitations under the License.
 #include "gtest/gtest.h"
 #include "milvus-storage/common/serde.h"
+#include <cstdint>
 
 namespace milvus_storage {
 
 class SerdeTest : public testing::Test {};
 
 TEST_F(SerdeTest, TestColumnOffsetSerde) {
-  std::vector<std::vector<int>> column_offsets = {{0, 1, 2}, {3, 4}, {5, 6, 7, 8}};
+  std::vector<std::vector<int64_t>> column_offsets = {{0, 1, 2}, {3, 4}, {5, 6, 7, 8}};
 
   std::string serialized = PackedMetaSerde::SerializeColumnOffsets(column_offsets);
   EXPECT_EQ(serialized, "0,1,2;3,4;5,6,7,8");
 
-  std::vector<std::vector<int>> deserialized = PackedMetaSerde::DeserializeColumnOffsets(serialized);
+  std::vector<std::vector<int64_t>> deserialized = PackedMetaSerde::DeserializeColumnOffsets(serialized);
   EXPECT_EQ(deserialized, column_offsets);
 
   // Test case: Empty input
-  std::vector<std::vector<int>> empty_offsets = {};
+  std::vector<std::vector<int64_t>> empty_offsets = {};
   serialized = PackedMetaSerde::SerializeColumnOffsets(empty_offsets);
   EXPECT_EQ(serialized, "");
 
@@ -36,7 +37,7 @@ TEST_F(SerdeTest, TestColumnOffsetSerde) {
   EXPECT_TRUE(deserialized.empty());
 
   // Test case: Single group
-  std::vector<std::vector<int>> single_group = {{1, 2, 3}};
+  std::vector<std::vector<int64_t>> single_group = {{1, 2, 3}};
   serialized = PackedMetaSerde::SerializeColumnOffsets(single_group);
   EXPECT_EQ(serialized, "1,2,3");
 
@@ -44,7 +45,7 @@ TEST_F(SerdeTest, TestColumnOffsetSerde) {
   EXPECT_EQ(deserialized, single_group);
 
   // Test case: Single column in each group
-  std::vector<std::vector<int>> single_column_groups = {{0}, {1}, {2}};
+  std::vector<std::vector<int64_t>> single_column_groups = {{0}, {1}, {2}};
   serialized = PackedMetaSerde::SerializeColumnOffsets(single_column_groups);
   EXPECT_EQ(serialized, "0;1;2");
